Authentication backend

+++ backend/ConsultingManager.Infra/Database/Models/CustomerTaskPoco.cs
@@ -9,7 +9,7 @@ namespace ConsultingManager.Infra.Database.Models
 
         public string Description { get; set; }
         public string Instructions { get; set; }
-        public long Duration { get; set; }
+        public int Duration { get; set; }
         public DateTime CreationDate { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EstimatedEndDate { get; set; }
